requirements.txt
torch
torchvision
Pillow
lxml
matplotlib
opencv-python
